grouped controllers for crud operations into multiple folders. One for each section of only summary controller uses a file instead of folder as it's just one get request.
